+++ core/moengage_flutter_android/android/src/main/kotlin/com/moengage/flutter/MoEngageFlutterPlugin.kt
@@ -11,15 +11,16 @@ import com.moengage.plugin.base.internal.PluginHelper
 import com.moengage.plugin.base.internal.setEventEmitter
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.FlutterPlugin.FlutterPluginBinding
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 
-class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
+class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler, ActivityAware {
 
     private val tag = "${MODULE_TAG}MoEngageFlutterPlugin"
-    private lateinit var channel: MethodChannel
     private lateinit var context: Context
     private val pluginHelper = PluginHelper()
 
@@ -32,14 +33,15 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
     override fun onAttachedToEngine(binding: FlutterPluginBinding) {
         Logger.print { "$tag onAttachedToEngine() : Registering MoEngageFlutterPlugin" }
         context = binding.applicationContext
-        initPlugin(binding.binaryMessenger)
+        if (methodChannel == null) {
+            initPlugin(binding.binaryMessenger)
+        }
     }
 
     override fun onDetachedFromEngine(binding: FlutterPluginBinding) {
         try {
             Logger.print { "$tag onDetachedFromEngine() : Registering MoEngageFlutterPlugin" }
             pluginHelper.onFrameworkDetached()
-            channel.setMethodCallHandler(null)
         } catch (t: Throwable) {
             Logger.print(LogLevel.ERROR, t) { "$tag onDetachedFromEngine() " }
         }
@@ -47,8 +49,9 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
 
     private fun initPlugin(binaryMessenger: BinaryMessenger) {
         try {
-            channel = MethodChannel(binaryMessenger, FLUTTER_PLUGIN_CHANNEL_NAME)
-            channel.setMethodCallHandler(this)
+            Logger.print { "$tag initPlugin(): Initializing MoEngage Flutter Plugin" }
+            methodChannel = MethodChannel(binaryMessenger, FLUTTER_PLUGIN_CHANNEL_NAME)
+            methodChannel?.setMethodCallHandler(this)
             setEventEmitter(EventEmitterImpl(::sendCallback))
             if (GlobalCache.lifecycleAwareCallbackEnabled){
                 Logger.print { "$tag initPlugin()  Adding App Background Listener: " }
@@ -63,7 +66,7 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
         try {
             Handler(Looper.getMainLooper()).post {
                 try {
-                    channel.invokeMethod(methodName, message)
+                    methodChannel?.invokeMethod(methodName, message)
                 } catch (t: Throwable) {
                     Logger.print(LogLevel.ERROR, t) { "$tag sendCallback() " }
                 }
@@ -369,5 +372,50 @@ class MoEngageFlutterPlugin : FlutterPlugin, MethodCallHandler {
         } catch (t: Throwable) {
             Logger.print(LogLevel.ERROR, t) { "$tag updatePushPermissionRequestCount() :" }
         }
+    }
+
+
+    /**
+     * Called when the plugin is attached to Flutter Activity.
+     * @param binding instance of [ActivityPluginBinding]
+     */
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        Logger.print { "$tag onAttachedToActivity() : Resetting methodChannel to `null`" }
+    }
+
+
+    /**
+     * Called when the plugin is Detached From Flutter Activity.
+     */
+    override fun onDetachedFromActivity() {
+        Logger.print { "$tag onDetachedFromActivity() : Resetting methodChannel to `null`" }
+        methodChannel = null
+    }
+
+    /**
+     * Called when the plugin is Detached From Flutter Activity for Config Changes
+     */
+    override fun onDetachedFromActivityForConfigChanges() {
+        Logger.print {
+            "$tag onDetachedFromActivityForConfigChanges() : Detached From Activity for Config changes"
+        }
+    }
+
+    /**
+     * Called when the plugin is Reattached to Flutter Activity For Config Changes.
+     * @param binding instance of [ActivityPluginBinding]
+     */
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        Logger.print {
+            "$tag onReattachedToActivityForConfigChanges() : ReAttached To Activity for Config changes"
+        }
+    }
+
+
+    companion object {
+        /**
+         * Static MethodChannel instance to avoid plugin reinitializing from Background Isolate
+         */
+        internal var methodChannel: MethodChannel? = null
     }
 }