+++ core/moengage_flutter_android/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MoEngage Flutter Android Plugin
 
+# Next Release Date
+
+## Next Version
+- Fix Android to Dart Method Channel Communication Breaking when `FirebaseMessaging.onBackgroundMessage` is used
+
 # 13-09-2023
 
 ## 1.0.0